More search components folder fixed
