+++ search.html
@@ -19,5 +19,6 @@
     <script type="module" src="functions.js"></script>
     <script src="header.js"></script>
     <script type="module" src="./search/search.js"></script>
+    <script type="module" src="./search/searchViewmodel.js"></script>
 </body>
 </html>
